requirements.txt
# IB-VTON Dependencies
# Core framework
agent-framework[azure]

# Web server
fastapi
uvicorn[standard]

# HTTP client
httpx

# Image processing
pillow

# Settings
pydantic
pydantic-settings

# Azure authentication
azure-identity

# Testing (dev)
pytest
pytest-asyncio
pytest-cov
